requirements.txt
asgiref==3.4.1
Django==3.2.7
django-ckeditor==6.1.0
django-debug-toolbar==3.2.2
django-js-asset==1.2.2
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
django-stubs==1.9.0
django-stubs-ext==0.3.1
mypy==0.910
mypy-extensions==0.4.3
Pillow==8.3.2
PyMySQL==1.0.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
types-pytz==2021.1.2
types-PyYAML==5.4.10
typing-extensions==3.10.0.2
